requirements.txt
aiomysql==0.2.0
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
asyncpg==0.30.0
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.15
exceptiongroup==1.2.2
Faker==34.0.2
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
kombu==5.4.2
Mako==1.3.8
MarkupSafe==2.1.5
mysqlclient==2.2.7
prompt_toolkit==3.0.48
pydantic==2.10.5
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==4.0.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.33.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.20.2
